# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Online Auction System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type AuctionStatus = (typeof AuctionStatus)[keyof typeof AuctionStatus];

export const AuctionStatus = {
  active: "active",
  ended: "ended",
  cancelled: "cancelled",
} as const;

export interface Auction {
  id: number;
  title: string;
  description: string;
  imageUrl?: string | null;
  startingPrice: number;
  currentPrice: number;
  category: string;
  endTime: string;
  sellerId: number;
  sellerUsername: string;
  status: AuctionStatus;
  bidCount: number;
  createdAt: string;
}

export interface Bid {
  id: number;
  auctionId: number;
  userId: number;
  username: string;
  amount: number;
  createdAt: string;
}

export interface AuctionDetail {
  auction: Auction;
  bids: Bid[];
  winner?: User | null;
}

export interface AuctionsListResponse {
  auctions: Auction[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateAuctionRequest {
  title: string;
  description: string;
  /** @minimum 0.01 */
  startingPrice: number;
  category: string;
  endTime: string;
}

export interface PlaceBidRequest {
  auctionId: number;
  amount: number;
}

export type DashboardResponseStats = {
  activeAuctions: number;
  totalBids: number;
  auctionsWon: number;
};

export interface DashboardResponse {
  myAuctions: Auction[];
  myBids: Bid[];
  wonAuctions: Auction[];
  stats: DashboardResponseStats;
}

export type GetAuctionsParams = {
  category?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type UploadAuctionImageBody = {
  image?: Blob;
};