# lib/api-zod/src/generated/types/auction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Online Auction System API
 * OpenAPI spec version: 0.1.0
 */
import type { AuctionStatus } from "./auctionStatus";

export interface Auction {
  id: number;
  title: string;
  description: string;
  imageUrl?: string | null;
  startingPrice: number;
  currentPrice: number;
  category: string;
  endTime: Date;
  sellerId: number;
  sellerUsername: string;
  status: AuctionStatus;
  bidCount: number;
  createdAt: Date;
}