# lib/api-zod/src/generated/types/auctionDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Online Auction System API
 * OpenAPI spec version: 0.1.0
 */
import type { Auction } from "./auction";
import type { Bid } from "./bid";
import type { User } from "./user";

export interface AuctionDetail {
  auction: Auction;
  bids: Bid[];
  winner?: User | null;
}