# lib/api-zod/src/generated/types/bid.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Online Auction System API
 * OpenAPI spec version: 0.1.0
 */

export interface Bid {
  id: number;
  auctionId: number;
  userId: number;
  username: string;
  amount: number;
  createdAt: Date;
}