# lib/api-zod/src/generated/types/createAuctionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Online Auction System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAuctionRequest {
  title: string;
  description: string;
  /** @minimum 0.01 */
  startingPrice: number;
  category: string;
  endTime: Date;
}