# lib/api-zod/src/generated/types/placeBidRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Online Auction System API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaceBidRequest {
  auctionId: number;
  amount: number;
}