# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Online Auction System API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  createdAt: Date;
}